LPD-94145 Extract the AI Assistant chat surface and panel header components
